refactor: decompose lexical editor
